code cleanup and refactor
