+++ dali-api/app/lib/delibs.ts
@@ -1,0 +1,3 @@
+// shared constants for delibs
+export const INITIAL_COLUMNS = ["No Decision", "Interview", "Reject"] as const;
+export const FINAL_COLUMNS = ["Accept", "Waitlist", "Reject"] as const;
